+++ lib/tfann.dart
@@ -9,4 +9,5 @@ library tfann;
 export 'src/network.dart';
 export 'src/train_set.dart';
 export 'src/activation_function.dart';
+export 'src/compile_network.dart';
 export 'src/linalg.dart' hide roundUp4, FRightMatrix;